+++ launch.sh
@@ -1,73 +1,102 @@
 #!/bin/bash
 
+# Function to detect the home directory
+detect_home_dir() {
+    case "$(uname -s)" in
+        Linux*|Darwin*) HOME_DIR="$HOME";;
+        CYGWIN*|MINGW*|MSYS*|MINGW32*|MINGW64*) HOME_DIR="C:/Users/$(whoami)";;
+        *) HOME_DIR="UNKNOWN"
+    esac
+}
+
 # Function to detect the operating system
 detect_os() {
     case "$(uname -s)" in
         Linux*)     OS=Linux;;
         Darwin*)    OS=Mac;;
-        CYGWIN*|MINGW*|MSYS*) OS=Windows;;
+        CYGWIN*|MINGW*|MSYS*|MINGW32*|MINGW64*) OS=Windows;;
         *)          OS="UNKNOWN"
     esac
-    echo "Detected OS: $OS"
+    echo "NCCA | Detected OS: $OS"
 }
 
-# Define the script directory
-SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
-cd $SCRIPT_DIR
-
-# Read the Python version from the .python-version file
-PYTHON_VERSION=$(cat "$SCRIPT_DIR/.python-version")
-
-echo "NCCA | Using Python version: $PYTHON_VERSION"
-
-# Function to install PyEnv on Linux
-install_pyenv_linux() {
-    if ! command -v pyenv &> /dev/null; then
-        echo "NCCA | PyEnv not found. Installing PyEnv..."
-        curl https://pyenv.run | bash
-        export PATH="$HOME/.pyenv/bin:$PATH"
-        eval "$(pyenv init --path)"
-        eval "$(pyenv init -)"
-        eval "$(pyenv virtualenv-init -)"
-    fi
+# Set the HOME environment variable to the detected home directory
+set_home_env() {
+    export HOME="$HOME_DIR"
 }
 
 # Function to install pyenv-win on Windows
 install_pyenv_windows() {
     if ! command -v pyenv &> /dev/null; then
         echo "NCCA | pyenv-win not found. Installing pyenv-win..."
-        git clone https://github.com/pyenv-win/pyenv-win.git "$HOME/.pyenv"
-        export PATH="$HOME/.pyenv/pyenv-win/bin:$HOME/.pyenv/pyenv-win/shims:$PATH"
+
+        INSTALL_DIR="$HOME/.pyenv"
+
+        # Clone pyenv-win repository
+        if [ -d "$INSTALL_DIR" ]; then
+            echo "NCCA | Removing existing $INSTALL_DIR directory..."
+            rm -rf "$INSTALL_DIR"
+        fi
+        
+        git clone https://github.com/pyenv-win/pyenv-win.git $INSTALL_DIR
+
+        # Add pyenv-win paths to the environment
+        export PATH="$INSTALL_DIR/pyenv-win/bin:$INSTALL_DIR/pyenv-win/shims:$PATH"
+
+        echo "NCCA | pyenv-win installed successfully."
+    else
+        echo "NCCA | pyenv-win already installed."
+    fi
+}
+
+# Function to verify pyenv installation
+verify_pyenv() {
+    if ! command -v pyenv &> /dev/null; then
+        echo "NCCA | Error: pyenv command not found. Exiting."
+        exit 1
+    else
+        echo "NCCA | pyenv command found."
     fi
 }
 
 # Main script execution
+# Main script execution
 main() {
     detect_os
+    detect_home_dir
+    set_home_env
 
     if [ "$OS" == "UNKNOWN" ]; then
         echo "NCCA | Unsupported OS. Exiting."
         exit 1
     fi
 
-    if [ "$OS" == "Linux" ]; then
-        install_pyenv_linux
-    elif [ "$OS" == "Windows" ]; then
+    if [ "$OS" == "Windows" ]; then
         install_pyenv_windows
+        verify_pyenv
     fi
+    
+    #eval "$(pyenv init --path)"
+    #eval "$(pyenv init -)"
+    #eval "$(pyenv virtualenv-init -)"
 
-    # Ensure PyEnv is in the PATH
-    export PATH="$HOME/.pyenv/bin:$PATH"
-    eval "$(pyenv init --path)"
-    eval "$(pyenv init -)"
-    eval "$(pyenv virtualenv-init -)"
+    # Define the script directory
+    SCRIPT_DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" && pwd )"
+    cd "$SCRIPT_DIR" || exit
+
+    # Read the Python version from the .python-version file
+    PYTHON_VERSION=$(<.python-version)
+    echo "NCCA | Using Python version: $PYTHON_VERSION"
 
     # Install or activate the specified Python version using PyEnv
     if pyenv versions | grep -q "$PYTHON_VERSION"; then
         echo "NCCA | Python $PYTHON_VERSION is already installed."
     else
         echo "NCCA | Python $PYTHON_VERSION is not installed. Installing..."
-        pyenv install "$PYTHON_VERSION"
+        pyenv install "$PYTHON_VERSION" || {
+            echo "NCCA | Error: Failed to install Python $PYTHON_VERSION."
+            exit 1
+        }
     fi
 
     pyenv local "$PYTHON_VERSION"
@@ -89,17 +118,20 @@ main() {
         source "$VENV_DIR/bin/activate"
     fi
 
-    python -m pip install --upgrade pip
-
-    requirements_file="$SCRIPT_DIR/requirements.txt"
-
-    if [ -f "$requirements_file" ]; then
-        echo "NCCA | Installing Requirements"
-        python -m pip install -r "$requirements_file"
+    # Check if requirements are already installed
+    if ! python -m pip freeze | grep -qF -f "$SCRIPT_DIR/requirements.txt"; then
+        # Requirements not installed, install them
+        requirements_file="$SCRIPT_DIR/requirements.txt"
+        if [ -f "$requirements_file" ]; then
+            echo "NCCA | Installing Requirements"
+            python -m pip install -r "$requirements_file"
+        else
+            echo "Requirements file not found: $requirements_file"
+            deactivate
+            exit 1
+        fi
     else
-        echo "Requirements file not found: $requirements_file"
-        deactivate
-        exit 1
+        echo "NCCA | Requirements already installed."
     fi
 
     main_script="$SCRIPT_DIR/NCCARenderFarm/main.py"
